feat: Add token login tasks 4.18-4.22

+++ osa_4/controllers/blogs.js
@@ -2,7 +2,8 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const logger = require('../utils/logger')
-
+const jwt = require('jsonwebtoken')
+const { tokenExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response, next) => {
   const blogs = await Blog.find({}).populate('user', {username:1, name:1})
@@ -20,9 +21,13 @@ blogsRouter.get('/', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next ) => {
   const body = request.body
+  console.log(request, "blogs roter tokens")
+  const user = request.user
+  try{
+  if(!tokenExtractor){
+    return response.status(401).json({error: "unauthorised to post"})
+  }
 
-  const user = await User.findById(body.userId)
-  
   if(!body.title || !body.url){
     return response.status(400).json({
       error: 'Content missing',
@@ -37,7 +42,7 @@ blogsRouter.post('/', async (request, response, next ) => {
     user: user._id
   })
   console.log(newBlog)
-  try{
+  
     const savedBlog = await newBlog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -53,12 +58,23 @@ blogsRouter.post('/', async (request, response, next ) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
  try{
-  const deleteBlog = await Blog.findByIdAndDelete(request.params.id)
+  console.log(request.params,"delete requeser")
+  const blog = await Blog.findById(request.params.id)
+  const userId = request.user.id
+  console.log(blog,"deleten blog olio")
+  if(blog.user.toString() === userId.toString()){
+    const deleteBlog = await Blog.findByIdAndDelete(request.params.id)
     if(deleteBlog){
       response.status(204).end()
     }else{
       response.status(404).end()
     }
+  } else{
+    response.status(401).json({
+      error: 'invalid token or user id'
+    })
+  }
+  
  }catch(error){
     next(error)
  }
